application: split voice and config use cases
